[kang] null 값에 따른 경로 삭제 정비 - vrtFilePath 에서 '' 값이 저장될 때, null 로 인식되지 않아 모든 파일이 날라가버리는 케이스가 생겼음. 이를 방지하기 위해 String null 확인 문장 추가.

+++ src/main/java/com/jiin/admin/website/view/component/MapVersionManagement.java
@@ -13,6 +13,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+import pl.jalokim.utils.string.StringUtils;
 
 import javax.annotation.Resource;
 import javax.transaction.Transactional;
@@ -49,7 +50,9 @@ public class MapVersionManagement {
                 mapVersionMapper.deleteById(version.getId());
                 String filePath = dataPath + version.getZipFilePath();
                 try {
-                    FileSystemUtil.deleteFile(filePath);
+                    if(!StringUtils.isBlank(filePath)) {
+                        FileSystemUtil.deleteFile(filePath);
+                    }
                 } catch (IOException e) {
                     log.error("ERROR - " + e.getMessage());
                 }
@@ -323,7 +326,9 @@ public class MapVersionManagement {
             if (mapVersionMapper.countLayersByMapVersionId(version.getId()) == 0) {
                 mapVersionMapper.deleteById(version.getId());
                 try {
-                    FileSystemUtil.deleteFile(dataPath + version.getZipFilePath());
+                    if(!StringUtils.isBlank(version.getZipFilePath())) {
+                        FileSystemUtil.deleteFile(dataPath + version.getZipFilePath());
+                    }
                 } catch (IOException e) {
                     log.error("ERROR - " + e.getMessage());
                 }
